Create 'createTask' method in TaskManager
When save button was clicked in 'CreateTask' class, it was hard to
understand what was going on. The logic was awful.

So, I have created separate 'createTask' method in TaskManager which
returns new task with id attached on it.

This is all for the readability of 'save' method in 'CreateTask'
class.